Improvements to sphinx
