+++ docs/sphinx/development/Platform-Detection.rst
