+++ docs/sphinx/development/Vulnerabilities-and-CHIPSEC-Modules.rst
@@ -1,8 +1,7 @@
 CHIPSEC Modules
 ===============
 
-Introduction
-------------
+A CHIPSEC module is just a python class that inherits from BaseModule and implements ``is_supported`` and ``run``. Modules are stored under the chipsec installation directory in a subdirectory "modules". The "modules" directory contains one subdirectory for each chipset that chipsec supports. There is also a directory for common modules that should apply to every platform.
 
     ===============================================  ================================================================================================
      ``chipsec/modules/``                            modules including tests or tools (that's where most of the chipsec functionality is)        
@@ -10,9 +9,7 @@ Introduction
      ``chipsec/modules/<platform>/``                 modules specific to <platform>                                                
      ``chipsec/modules/tools/``                      security tools based on CHIPSEC framework (fuzzers, etc.)                                   
     ===============================================  ================================================================================================
-
-A CHIPSEC module is just a python class that inherits from BaseModule and implements ``is_supported`` and ``run``. Modules are stored under the chipsec installation directory in a subdirectory "modules". The "modules" directory contains one subdirectory for each chipset that chipsec supports. There is also a directory for common modules that should apply to every platform.
-
+    
 Internally the chipsec application uses the concept of a module name, which is a string of the form: ``common.bios_wp``.
 This means module ``common.bios_wp`` is a python script called ``bios_wp.py`` that is stored at ``<ROOT_DIR>\chipsec\modules\common\``.
 
@@ -288,4 +285,4 @@ Modules
 
 .. toctree:: 
 
-    List of modules <modules/chipsec.modules.rst>
+    List of modules <../modules/chipsec.modules.rst>
