+++ docs/sphinx/installation/Install in DAL Win.rst
