+++ docs/sphinx/installation/Install in MacOS.rst
