+++ docs/sphinx/installation/Install in Windows.rst
