+++ docs/sphinx/installation/USB with UEFI Shell.rst
