+++ docs/sphinx/start/Download.rst
@@ -6,7 +6,7 @@ GitHub repository
 
 CHIPSEC source files are maintained in a GitHub repository:
 
-.. image:: _images/GHRepo.png
+.. image:: ../_images/GHRepo.png
    :target: https://github.com/chipsec/chipsec
 
 Releases
@@ -14,9 +14,16 @@ Releases
 
 You can find the latest release here:
 
-.. image:: _images/LR.png
+.. image:: ../_images/LR.png
    :target: https://github.com/chipsec/chipsec/releases/latest
 
 Older releases can be found `here <https://github.com/chipsec/chipsec/releases>`_
 
 After downloading there are some steps to follow to build the driver and run, please refer to :ref:`Installation <Installing-Chipsec>` and :ref:`running CHIPSEC <Running-Chipsec>`
+
+Python
+------
+
+Python downloads:
+
+https://www.python.org/downloads/
